reorganize lookup functionality and add roles context
