+++ src/console-runners/seed-data.runner.ts
@@ -1,6 +1,6 @@
 import { dbServer } from '#core/servers/db.server.js';
 import { db } from '#dals/mock.data.js';
-import { getUnidadesProponentesContext } from '#dals/unidad-proponente/index.js';
+import { getRolesContext, getUnidadesProponentesContext } from '#dals/lookup/index.js';
 import { getUserContext } from '#dals/user/index.js';
 
 export const run = async (connectionString: string) => {
@@ -10,6 +10,8 @@ export const run = async (connectionString: string) => {
     await getUserContext().insertMany(db.usuarios);
 
     await getUnidadesProponentesContext().insertMany(db.unidadProponentes);
+
+    await getRolesContext().insertMany(db.roles);
 
     console.log('Data seeded successfully');
