+++ src/dals/lookup/role.context.ts
@@ -1,0 +1,4 @@
+import { Lookup } from '#common/models/lookup.js';
+import { dbServer } from '#core/servers/index.js';
+
+export const getRolesContext = () => dbServer.db?.collection<Lookup>('roles');
