--- src/dals/unidad-proponente/index.ts
@@ -1,2 +1,0 @@
-export * from './unidad-proponente.repository.js';
-export * from './unidad-proponente.context.js';
