+++ src/dals/user/user.repository.ts
@@ -5,10 +5,16 @@ import { Usuario } from './user.model.js';
 
 export const userRepository = {
   getUserList: async (page?: number, pageSize?: number): Promise<Usuario[]> => {
-    const skip = Boolean(page) ? (page - 1) * pageSize : 0;
-    const limit = pageSize ?? 0;
+    const existPageAndPageSize = page !== undefined && page !== null && pageSize !== undefined && pageSize !== null;
 
-    return await getUserContext().find().skip(skip).limit(limit).toArray();
+    if (existPageAndPageSize) {
+      const skip = page !== undefined && page !== null ? page * pageSize : 0;
+      const limit = pageSize ?? 0;
+
+      return await getUserContext().find().skip(skip).limit(limit).toArray();
+    }
+
+    return await getUserContext().find().toArray();
   },
   getUser: async (id: string, selectedFields: SelectedFields<Usuario>) =>
     await getUserContext().findOne({ _id: mapStringToObjectId(id) }, { projection: selectedFields }),
