+++ src/index.ts
@@ -5,13 +5,13 @@ import { ENV } from '#core/constants/index.js';
 import { logger } from '#core/logger/index.js';
 import { createRestApiServer, dbServer } from '#core/servers/index.js';
 import { userApi } from '#pods/user/index.js';
-import { lookupApi } from '#pods/unidad-proponente/index.js';
+import { lookupApi } from '#pods/lookup/index.js';
 
 const app = createRestApiServer();
 
 app.use(logRequestMiddleware(logger));
 
-app.use('/api/user', userApi);
+app.use('/api/usuario', userApi);
 
 app.use('/api/lookup', lookupApi);
